Refactor ServiceBus messagehandling to new class
